scraper/requirements.txt
pymongo
playwright
python-dotenv